Gracia Final support 4/6
git-svn-id: http://svn.l2jserver.com/trunk@2948 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/network/clientpackets/RequestExEnchantSkillInfoDetail.java
@@ -64,7 +64,7 @@ public final class RequestExEnchantSkillInfoDetail extends L2GameClientPacket
             return;
         
         int bookId = 0;
-        int reqCount = 0;
+        long reqCount = 0;
         // require book for first level
         int enchantLevel = _skillLvl%100;
         // if going to first level OR going to Original level(untraining)
